feat: remove version parsing from #name selector
BREAKING CHANGE: You can no longer select a version along with an name
selector (i.e. `#npm@9.0.0`).  Use `#name:semver()` instead.

semver and dependency type selectors collide in syntax, so we are no
longer trying to parse a version with a name selector.

An example can show why this is impossible to parse:
`@npmcli/test-package@1.2.3-pre.prod`.

There is no way to know if `.prod` is supposed to be the `.prod`
dependency type selector or if it is part of the pre-release section of
the version.  So now you must use one of these two syntaxes depending on
what the version actually is:

If `.prod` is part of the version:
`#@npmcli/test-package:semver(1.2.3-pre.prod)`

If `.prod` is a dependency type selector:
`#@npmcli/test-package:semver(1.2.3-pre).prod`